+++ Model/EstusFlask.h
@@ -7,7 +7,7 @@
 class EstusFlask: public Item {
 public:
     explicit EstusFlask(std::string itemName = "Fiaschetta di Cura",int itemPrice = 0,
-                        std::string itemDescriptin = "Fiaschette che curano 30hp", int maxItemCount = 5, int itemCount = 5, float x = 412, float y = 2, 
+                        std::string itemDescriptin = "Fiaschette che curano 30hp", int maxItemCount = 5, int itemCount = 5, int x = 412, int y = 2, 
                         int width = 20, int height = 21);
     ~EstusFlask();
     //TODO void use(Hero& hero);
